pdf_merge_split_ocr/requirements.txt
pypdf>=4.0
pdfplumber>=0.10
pytesseract>=0.3
Pillow>=10.0
